Moving NotFound to common
