+++ src/containers/common/PublicLayout.js
@@ -3,7 +3,7 @@ import { Nav, NavItem, Navbar } from "react-bootstrap";
 import { LinkContainer } from 'react-router-bootstrap';
 import { Link, Route, Switch } from "react-router-dom";
 import Home from "../Home";
-import NotFound from "../NotFound";
+import NotFound from "./NotFound";
 
 export default class PublicLayout extends Component {
     render() {
